user with admin & teacher

+++ routes/web.php
@@ -10,6 +10,8 @@ use App\Http\Controllers\Schedule\ScheduleController;
 use App\Http\Controllers\Student\StudentController;
 use App\Http\Controllers\Subject\SubjectController;
 use App\Http\Controllers\Teacher\TeacherController;
+use App\Http\Middleware\AdminMiddleware;
+use App\Http\Middleware\TeacherMiddleware;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', function () {
@@ -24,55 +26,57 @@ Route::middleware('auth')->group(function () {
 
 Route::middleware(['auth', 'verified'])->group(function () {
 
-    // Dashboard
-    Route::prefix('dashboard')->group(function () {
-        Route::get('/', [DashboardController::class, 'index'])->name('dashboard');
-    });
+    // Only admin routes
+    Route::middleware([AdminMiddleware::class])->group(function () {
 
-    // Faculties
-    Route::prefix('faculties')->group(function () {
-        Route::get('/{id}', [FacultyController::class, 'show'])->name('faculties.show');
-        Route::get('/', [FacultyController::class, 'getFaculties'])->name('faculties.index');
-        Route::post('/', [FacultyController::class, 'createFaculty'])->name('faculties.create');
-        Route::delete('/{id}', [FacultyController::class, 'deleteFaculty'])->name('faculties.delete');
-    });
+        // Dashboard
+        Route::prefix('dashboard')->group(function () {
+            Route::get('/', [DashboardController::class, 'index'])->name('dashboard');
+        });
 
-    // Departments
-    Route::prefix('departments')->group(function () {
-        Route::get('/{id}', [DepartmentController::class, 'show'])->name('departments.show');
-        Route::post('/{facultyId}', [DepartmentController::class, 'createDepartment'])->name('departments.create');
-        Route::delete('/{id}', [DepartmentController::class, 'deleteDepartment'])->name('departments.delete');
+        // Faculties
+        Route::prefix('faculties')->group(function () {
+            Route::get('/', [FacultyController::class, 'getFaculties'])->name('faculties.index');
+            Route::get('/{id}', [FacultyController::class, 'show'])->name('faculties.show');
+            Route::post('/', [FacultyController::class, 'createFaculty'])->name('faculties.create');
+            Route::delete('/{id}', [FacultyController::class, 'deleteFaculty'])->name('faculties.delete');
+        });
+    
+        // Departments
+        Route::prefix('departments')->group(function () {
+            Route::get('/{id}', [DepartmentController::class, 'show'])->name('departments.show');
+            Route::post('/{facultyId}', [DepartmentController::class, 'createDepartment'])->name('departments.create');
+            Route::delete('/{id}', [DepartmentController::class, 'deleteDepartment'])->name('departments.delete');
+        });
+
+        // Students
+        Route::prefix('students')->group(function () {
+            Route::post('/{groupId}', [StudentController::class, 'createStudent'])->name('students.create');
+            Route::delete('/{id}', [StudentController::class, 'deleteStudent'])->name('students.delete');
+        });
+
+        // Subjects
+        Route::prefix('subjects')->group(function () {
+            Route::post('/{groupId}', [SubjectController::class, 'createSubject'])->name('subjects.create');
+            Route::delete('/{id}', [SubjectController::class, 'deleteSubject'])->name('subjects.delete');
+        });
     });
 
     // Groups
     Route::prefix('groups')->group(function () {
         Route::get('/')->name('groups');
-        Route::get('/{id}', [GroupController::class, 'show'])->name('groups.show');
-        Route::post('/{departmentId}', [GroupController::class, 'createGroup'])->name('groups.create');
-        Route::delete('/{id}', [GroupController::class, 'deleteGroup'])->name('groups.delete');
+        Route::get('/{id}', [GroupController::class, 'show'])->name('groups.show')->middleware([AdminMiddleware::class]);
+        Route::post('/{departmentId}', [GroupController::class, 'createGroup'])->name('groups.create')->middleware([AdminMiddleware::class]);
+        Route::delete('/{id}', [GroupController::class, 'deleteGroup'])->name('groups.delete')->middleware([AdminMiddleware::class]);
         Route::get('/{id}/subjects', [GroupController::class, 'getGroupSubjects'])->name('groups.subjects');
-        Route::post('/{id}/assign-subjects', [GroupController::class, 'assignSubjectsToGroup'])->name('groups.assignSubjects');
     });
-
-    // Students
-    Route::prefix('students')->group(function () {
-        Route::post('/{groupId}', [StudentController::class, 'createStudent'])->name('students.create');
-        Route::delete('/{id}', [StudentController::class, 'deleteStudent'])->name('students.delete');
-    });
-
-    // Subjects
-    Route::prefix('subjects')->group(function () {
-        Route::post('/{groupId}', [SubjectController::class, 'createSubject'])->name('subjects.create');
-        Route::delete('/{id}', [SubjectController::class, 'deleteSubject'])->name('subjects.delete');
-    });
-
 
     // Teacher
     Route::prefix('teacher')->group(function () {
         Route::get('/', [TeacherController::class, 'index'])->name('teacher.index');
     });
     
-    // Teacher
+    // Schedule
     Route::prefix('schedule')->group(function () {
         Route::get('/', [ScheduleController::class, 'index'])->name('schedule.index');
         Route::post('/store', [ScheduleController::class, 'store'])->name('schedule.store');
@@ -85,7 +89,6 @@ Route::middleware(['auth', 'verified'])->group(function () {
         Route::get('/thursday', [ScheduleController::class, 'thursday'])->name('schedule.thursday');
         Route::get('/friday', [ScheduleController::class, 'friday'])->name('schedule.friday');
         Route::get('/saturday', [ScheduleController::class, 'saturday'])->name('schedule.saturday');
-
     });
 
     // Attendance
